+++ bootProject/src/main/java/com/busanit501/bootproject/repository/UserRepostiory.java
@@ -3,5 +3,5 @@ package com.busanit501.bootproject.repository;
 import com.busanit501.bootproject.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepostiory extends JpaRepository<User, Integer> {
+public interface UserRepostiory extends JpaRepository<User, Long> {
 }
